Fixup all native diagnostics under the following clang options: -Weverything -Werror -pedantic -pedantic-errors -Wno-c++98-compat -Wno-c++98-compat-pedantic -Wno-padded -Wno-ctad-maybe-unsupported -Wno-documentation-unknown-command -Wno-exit-time-destructors -Wno-unsafe-buffer-usage -Wno-global-constructors -Wno-switch-enum -Wno-old-style-cast

+++ Coral.Native/Source/Coral/Assembly.cpp
@@ -59,7 +59,7 @@ namespace Coral {
 
 			int32_t typeCount = 0;
 			s_ManagedFunctions.GetAssemblyTypesFptr(result.m_AssemblyID, nullptr, &typeCount);
-			std::vector<TypeId> typeIds(typeCount);
+			std::vector<TypeId> typeIds(static_cast<size_t>(typeCount));
 			s_ManagedFunctions.GetAssemblyTypesFptr(result.m_AssemblyID, typeIds.data(), &typeCount);
 
 			for (auto typeId : typeIds)
@@ -90,7 +90,7 @@ namespace Coral {
 
 			int32_t typeCount = 0;
 			s_ManagedFunctions.GetAssemblyTypesFptr(result.m_AssemblyID, nullptr, &typeCount);
-			std::vector<TypeId> typeIds(typeCount);
+			std::vector<TypeId> typeIds(static_cast<size_t>(typeCount));
 			s_ManagedFunctions.GetAssemblyTypesFptr(result.m_AssemblyID, typeIds.data(), &typeCount);
 
 			for (auto typeId : typeIds)
