+++ Coral.Native/Source/Coral/Type.cpp
@@ -35,11 +35,11 @@ namespace Coral {
 			s_ManagedFunctions.GetInterfaceTypeCountFptr(m_Id, &count);
 
 			std::vector<TypeId> typeIds;
-			typeIds.resize(count);
+			typeIds.resize(static_cast<size_t>(count));
 			s_ManagedFunctions.GetInterfaceTypesFptr(m_Id, typeIds.data());
 
 			m_InterfaceTypes = std::vector<Type*>();
-			m_InterfaceTypes->reserve(count);
+			m_InterfaceTypes->reserve(static_cast<size_t>(count));
 
 			for (auto id : typeIds)
 			{
@@ -76,7 +76,7 @@ namespace Coral {
 	{
 		int32_t methodCount = 0;
 		s_ManagedFunctions.GetTypeMethodsFptr(m_Id, nullptr, &methodCount);
-		std::vector<ManagedHandle> handles(methodCount);
+		std::vector<ManagedHandle> handles(static_cast<size_t>(methodCount));
 		s_ManagedFunctions.GetTypeMethodsFptr(m_Id, handles.data(), &methodCount);
 
 		std::vector<MethodInfo> methods(handles.size());
@@ -90,7 +90,7 @@ namespace Coral {
 	{
 		int32_t fieldCount = 0;
 		s_ManagedFunctions.GetTypeFieldsFptr(m_Id, nullptr, &fieldCount);
-		std::vector<ManagedHandle> handles(fieldCount);
+		std::vector<ManagedHandle> handles(static_cast<size_t>(fieldCount));
 		s_ManagedFunctions.GetTypeFieldsFptr(m_Id, handles.data(), &fieldCount);
 
 		std::vector<FieldInfo> fields(handles.size());
@@ -104,7 +104,7 @@ namespace Coral {
 	{
 		int32_t propertyCount = 0;
 		s_ManagedFunctions.GetTypePropertiesFptr(m_Id, nullptr, &propertyCount);
-		std::vector<ManagedHandle> handles(propertyCount);
+		std::vector<ManagedHandle> handles(static_cast<size_t>(propertyCount));
 		s_ManagedFunctions.GetTypePropertiesFptr(m_Id, handles.data(), &propertyCount);
 
 		std::vector<PropertyInfo> properties(handles.size());
@@ -123,7 +123,7 @@ namespace Coral {
 	{
 		int32_t attributeCount;
 		s_ManagedFunctions.GetTypeAttributesFptr(m_Id, nullptr, &attributeCount);
-		std::vector<ManagedHandle> attributeHandles(attributeCount);
+		std::vector<ManagedHandle> attributeHandles(static_cast<size_t>(attributeCount));
 		s_ManagedFunctions.GetTypeAttributesFptr(m_Id, attributeHandles.data(), &attributeCount);
 
 		std::vector<Attribute> result(attributeHandles.size());
